requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
jinja2==3.1.2
pyyaml==6.0.1
python-multipart==0.0.6
networkx==3.2.1
scipy==1.11.4
numpy==1.24.3
